Created new shader for phong lighting, bugfixing
- New shader currently just visualises different textures (albedo, spec and normal)
- Updated CMakeLIsts
- Added new push constants types to control which texture is applied}

TODO:
- Add controls for controlling light direction CPU-side
- Add phong lighting calculations to fragment shader
- Add descriptor layout creation logic to model class so models can create their own descriptor sets for textures (currently loading textures as separate members in renderer)

+++ Cassidy/Utils/DescriptorBuilder.cpp
@@ -176,7 +176,7 @@ VkDescriptorSetLayout cassidy::DescriptorLayoutCache::createDescLayout(VkDescrip
     layoutInfo.bindings.push_back(layoutCreateInfo->pBindings[i]);
 
     // Keep track of whether bindings are in ascending order or not:
-    if (layoutCreateInfo->pBindings[i].binding > lastBinding)
+    if (static_cast<int>(layoutCreateInfo->pBindings[i].binding) > lastBinding)
       lastBinding = layoutCreateInfo->pBindings[i].binding;
     else
       isSorted = false;
